+++ src/App.tsx
@@ -6,18 +6,18 @@ import { AuthProvider } from './contexts/AuthContext';
 import Header from './components/common/Header';
 import Footer from './components/common/Footer';
 import Homepage from './pages/Homepage';
-import MyReviewsPage from './pages/MyReviews';
+import MyReviewsPage from './pages/user/MyReviews';
 import EntitiesDirectory from './pages/EntitiesDirectory';
 import EntityProfile from './pages/EntityProfile';
-import WriteReview from './pages/WriteReview';
+import WriteReview from './pages/user/WriteReview';
 import Reviews from './pages/Reviews';
 import Auth from './pages/Auth';
-import ProfileSettings from './pages/ProfileSettings';
-import EntityRegistration from './pages/EntityRegistration';
-import EntityRegistrationSuccess from './pages/EntityRegistrationSuccess';
+import ProfileSettings from './pages/user/ProfileSettings';
+import EntityRegistration from './pages/admin/EntityRegistration';
+import EntityRegistrationSuccess from './pages/admin/EntityRegistrationSuccess';
 import SearchResults from './pages/SearchResults';
-import SuperAdminDashboard from './pages/SuperAdminDashboard';
-import TestScrollPage from './pages/test/TestScrollPage';
+import SuperAdminDashboard from './pages/super_admin/SuperAdminDashboard';
+import TestScrollPage from './test/TestScrollPage';
 import NotFound from './pages/NotFound';
 import { useAuth } from '@/contexts/AuthContext';
 import { useUserRoles } from '@/hooks/useUserRoles';
