+++ src/pages/admin/EntityRegistration.tsx
