+++ src/pages/admin/EntityRegistrationSuccess.tsx
